# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Flick API
  version: 0.1.0
  description: HTTP contract for Flick web clients.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      operationId: getHealthz
      responses:
        "200":
          description: API process is alive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /readyz:
    get:
      operationId: getReadyz
      responses:
        "200":
          description: API dependencies are ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: API dependency is not ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /metrics:
    get:
      operationId: getMetrics
      responses:
        "200":
          description: Prometheus metrics.
          content:
            text/plain:
              schema:
                type: string
  /api/secrets:
    post:
      operationId: createSecret
      summary: Create an encrypted secret.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSecretRequest"
      responses:
        "201":
          description: Secret created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSecretResponse"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Payload too large.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/secrets/{id}:
    get:
      operationId: getSecretMetadata
      summary: Fetch safe metadata needed to derive an access proof.
      parameters:
        - $ref: "#/components/parameters/SecretId"
      responses:
        "200":
          description: Metadata found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSecretMetadataResponse"
        "404":
          description: Secret missing, expired, or consumed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: Secret already consumed or expired.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/secrets/{id}/open:
    post:
      operationId: openSecret
      summary: Atomically verify access, consume the secret, and return encrypted payload.
      parameters:
        - $ref: "#/components/parameters/SecretId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenSecretRequest"
      responses:
        "200":
          description: Secret opened.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenSecretResponse"
        "404":
          description: Secret missing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access proof is invalid. The fifth invalid proof attempt removes the stored payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: Secret already consumed or expired.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /internal/secrets/{id}/cleanup:
    post:
      operationId: cleanupSecret
      summary: Internal worker endpoint for idempotent cleanup.
      security:
        - internalToken: []
      parameters:
        - $ref: "#/components/parameters/SecretId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CleanupSecretRequest"
      responses:
        "200":
          description: Cleanup completed or was already complete.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CleanupSecretResponse"
        "400":
          description: Invalid cleanup metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid internal token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    internalToken:
      type: apiKey
      in: header
      name: X-Flick-Internal-Token
  parameters:
    SecretId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    CreateSecretRequest:
      type: object
      additionalProperties: false
      required:
        - kind
        - ciphertext
        - nonce
        - kdf
        - access
        - size_bytes
        - ttl_seconds
      properties:
        kind:
          type: string
          enum: [text, file]
        ciphertext:
          type: string
          description: Base64-encoded encrypted payload.
        nonce:
          type: string
          description: Base64-encoded AES-GCM nonce.
        kdf:
          $ref: "#/components/schemas/KdfParams"
        access:
          $ref: "#/components/schemas/AccessVerifier"
        encrypted_filename:
          type: string
          description: Required for file secrets. Browser-encrypted JSON envelope containing filename metadata.
        content_type:
          type: string
          description: Media type for file download, or application/octet-stream.
        size_bytes:
          type: integer
          minimum: 0
          description: Plain payload byte length before browser encryption.
        ttl_seconds:
          type: integer
          enum: [600, 3600, 86400]
        max_views:
          type: integer
          minimum: 1
          default: 1
    CreateSecretResponse:
      type: object
      additionalProperties: false
      required: [id, expires_at]
      properties:
        id:
          type: string
        expires_at:
          type: string
          format: date-time
    AccessVerifier:
      type: object
      additionalProperties: false
      required: [kdf, proof]
      properties:
        kdf:
          $ref: "#/components/schemas/KdfParams"
        proof:
          type: string
          description: Base64-encoded client-derived access proof.
    AccessMetadata:
      type: object
      additionalProperties: false
      required: [kdf]
      properties:
        kdf:
          $ref: "#/components/schemas/KdfParams"
    GetSecretMetadataResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - access
        - size_bytes
        - expires_at
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [text, file]
        access:
          $ref: "#/components/schemas/AccessMetadata"
        size_bytes:
          type: integer
          minimum: 0
        expires_at:
          type: string
          format: date-time
    OpenSecretRequest:
      type: object
      additionalProperties: false
      required: [access_proof]
      properties:
        access_proof:
          type: string
          description: Base64-encoded proof derived from local user input and metadata access KDF.
    OpenSecretResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - ciphertext
        - nonce
        - kdf
        - size_bytes
        - expires_at
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [text, file]
        ciphertext:
          type: string
          description: Base64-encoded encrypted payload.
        nonce:
          type: string
          description: Base64-encoded AES-GCM nonce.
        kdf:
          $ref: "#/components/schemas/KdfParams"
        encrypted_filename:
          type: string
          description: Browser-encrypted JSON envelope containing filename metadata.
        content_type:
          type: string
          description: Media type for file download, or application/octet-stream.
        size_bytes:
          type: integer
          minimum: 0
        expires_at:
          type: string
          format: date-time
    KdfParams:
      type: object
      additionalProperties: false
      required:
        - algorithm
        - salt
        - iterations
        - key_length_bits
      properties:
        algorithm:
          type: string
          enum: [PBKDF2-SHA-256]
        salt:
          type: string
          description: Base64-encoded random KDF salt.
        iterations:
          type: integer
          minimum: 600000
        key_length_bits:
          type: integer
          enum: [256]
    CleanupSecretRequest:
      type: object
      additionalProperties: false
      required: [job_id, reason]
      properties:
        job_id:
          type: string
          minLength: 1
        reason:
          type: string
          enum: [consumed, expired, orphan, manual, retry]
    CleanupSecretResponse:
      type: object
      additionalProperties: false
      required: [id, cleaned]
      properties:
        id:
          type: string
        cleaned:
          type: boolean